Create a lnading page
Related to #1
